reorganise modal files
