+++ jsapp/js/project/FormLanguagesManager/index.ts
@@ -1,0 +1,3 @@
+export { default as FormLanguagesManager } from './FormLanguagesManager'
+export * from './FormLanguagesManager'
+export * from './openFormLanguagesModal'
